+++ crates/core/src/models.rs
@@ -33,7 +33,7 @@ pub struct Model {
     pub tpd_limit: Option<u64>,
     pub tpm_limit_month: Option<u64>,
     pub rps_limit: Option<f32>,
-    pub quality_score: Option<f32>,
+    pub quality_score: Option<f64>,
     pub avg_latency_ms: Option<u32>,
     pub is_enabled: bool,
     pub notes: Option<String>,
